refactor(-): :hammer: move fixtures logs dietas
